destructure folders
