+++ fixtures/simple.js
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export default () => <div>This file exports a very simple component</div>;
