+++ fixtures/withAbsoluteImport.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import Resolver from 'react-node-resolver';
+
+export default () => (
+  <div>
+    <p>This file tests an absolute import</p>
+  </div>
+);
