+++ fixtures/withRelativeImport.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import Simple from './Simple';
+
+export default () => (
+  <div>
+    <p>Here we are importing a simple component and rendering it:</p>
+    <Simple />
+  </div>
+);
