+++ src/codesandboxer/constants.js
@@ -41,3 +41,11 @@ export const newpkgJSON = (
       .join(',\n    ')}
   }
 }`;
+
+export const codesandboxURL =
+  'https://codesandbox.io/api/v1/sandboxes/define?query=module=/example.js';
+export const codesandboxURLJSON =
+  'https://codesandbox.io/api/v1/sandboxes/define?json=1';
+
+export const getSandboxUrl = (id: string, type?: string = 's') =>
+  `https://codesandbox.io/${type}/${id}?module=/example.js`;
