+++ src/codesandboxer/parseFile/index.js
