+++ src/codesandboxer/utils/getAllImports.js
@@ -4,7 +4,8 @@ import type { Import } from '../types';
 
 const getAllImports = (code: string): Array<Import> => {
   let imports = code.match(new RegExp(importPattern, 'g'));
-  if (imports) return imports.map(mpt => new RegExp(importPattern).exec(mpt));
+  if (imports)
+    return imports.map(mpt => new RegExp(importPattern).exec(mpt).map(a => a));
   else return [];
 };
