+++ src/index.js
@@ -1,6 +1,8 @@
 // @flow
-export { default as parseFile } from './parseFile';
-export { default as replaceImports } from './replaceImports';
-export { default as fetchRelativeFile } from './fetchRelativeFile';
+export { default as parseFile } from './codesandboxer/parseFile';
+export { default as replaceImports } from './codesandboxer/replaceImports';
+export {
+  default as fetchRelativeFile,
+} from './codesandboxer/fetchRelativeFile';
 
-export { default } from './CodeSandboxer';
+export { default } from './react-codesandboxer/CodeSandboxer';
